move building images to data/
