+++ data/building-images.js
@@ -1,0 +1,28 @@
+// @flow
+export const buildingImages = {
+  'alumni-hall': require('./images/buildinghours/w1440/alumni-hall.jpg'),
+  'boe': require('./images/buildinghours/w1440/boe.jpg'),
+  'bookstore': require('./images/buildinghours/w1440/bookstore.jpg'),
+  'buntrock': require('./images/buildinghours/w1440/buntrock.jpg'),
+  'cage': require('./images/buildinghours/w1440/cage.jpg'),
+  'christiansen': require('./images/buildinghours/w1440/christiansen.jpg'),
+  'convenience': require('./images/buildinghours/w1440/convenience.jpg'),
+  'disco': require('./images/buildinghours/w1440/disco.jpg'),
+  'dittmann': require('./images/buildinghours/w1440/dittmann.jpg'),
+  'hall-of-music': require('./images/buildinghours/w1440/hall-of-music.jpg'),
+  'halvorson': require('./images/buildinghours/w1440/halvorson.jpg'),
+  'old-main': require('./images/buildinghours/w1440/old-main.jpg'),
+  'pause-kitchen': require('./images/buildinghours/w1440/pause-kitchen.jpg'),
+  'print': require('./images/buildinghours/w1440/print.jpg'),
+  'post-office': require('./images/buildinghours/w1440/post-office.jpg'),
+  'regents-hall': require('./images/buildinghours/w1440/regents-hall.jpg'),
+  'regents-math': require('./images/buildinghours/w1440/regents-math.jpg'),
+  'rolvaag-library': require('./images/buildinghours/w1440/rolvaag-library.jpg'),
+  'skifter-studioa': require('./images/buildinghours/w1440/skifter-studioa.jpg'),
+  'skoglund': require('./images/buildinghours/w1440/skoglund.jpg'),
+  'stav': require('./images/buildinghours/w1440/stav.jpg'),
+  'theater': require('./images/buildinghours/w1440/theater.jpg'),
+  'tom-porter': require('./images/buildinghours/w1440/tom-porter.jpg'),
+  'tomson': require('./images/buildinghours/w1440/tomson.jpg'),
+  'wellness': require('./images/buildinghours/w1440/wellness.jpg'),
+}
