--- views/building-hours/images.js
@@ -1,28 +1,0 @@
-// @flow
-export const buildingImages = {
-  'alumni-hall': require('../../data/images/buildinghours/w1440/alumni-hall.jpg'),
-  'boe': require('../../data/images/buildinghours/w1440/boe.jpg'),
-  'bookstore': require('../../data/images/buildinghours/w1440/bookstore.jpg'),
-  'buntrock': require('../../data/images/buildinghours/w1440/buntrock.jpg'),
-  'cage': require('../../data/images/buildinghours/w1440/cage.jpg'),
-  'christiansen': require('../../data/images/buildinghours/w1440/christiansen.jpg'),
-  'convenience': require('../../data/images/buildinghours/w1440/convenience.jpg'),
-  'disco': require('../../data/images/buildinghours/w1440/disco.jpg'),
-  'dittmann': require('../../data/images/buildinghours/w1440/dittmann.jpg'),
-  'hall-of-music': require('../../data/images/buildinghours/w1440/hall-of-music.jpg'),
-  'halvorson': require('../../data/images/buildinghours/w1440/halvorson.jpg'),
-  'old-main': require('../../data/images/buildinghours/w1440/old-main.jpg'),
-  'pause-kitchen': require('../../data/images/buildinghours/w1440/pause-kitchen.jpg'),
-  'print': require('../../data/images/buildinghours/w1440/print.jpg'),
-  'post-office': require('../../data/images/buildinghours/w1440/post-office.jpg'),
-  'regents-hall': require('../../data/images/buildinghours/w1440/regents-hall.jpg'),
-  'regents-math': require('../../data/images/buildinghours/w1440/regents-math.jpg'),
-  'rolvaag-library': require('../../data/images/buildinghours/w1440/rolvaag-library.jpg'),
-  'skifter-studioa': require('../../data/images/buildinghours/w1440/skifter-studioa.jpg'),
-  'skoglund': require('../../data/images/buildinghours/w1440/skoglund.jpg'),
-  'stav': require('../../data/images/buildinghours/w1440/stav.jpg'),
-  'theater': require('../../data/images/buildinghours/w1440/theater.jpg'),
-  'tom-porter': require('../../data/images/buildinghours/w1440/tom-porter.jpg'),
-  'tomson': require('../../data/images/buildinghours/w1440/tomson.jpg'),
-  'wellness': require('../../data/images/buildinghours/w1440/wellness.jpg'),
-}
